% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ite.R
\docType{data}
\name{Y_poisson}
\alias{Y_poisson}
\title{A. Thaliana Simulated Phenotype matrix}
\format{
## `Y_poisson`
A vector with 328 observations corresponding to the 328 ecotypes.
}
\usage{
Y_poisson
}
\description{
This is a phenotype matrix simulated from 9,000 SNPs. SNPs at positions 450, 1350, 2250, 3150, 4050,
4950, 5850, 6750, 7650, and 8550 have nonzero coefficients. Further, the data was simulated under a poisson mixed effects model
with both a kinship random effect and an overdispersion random effect. The data was simulated using the kinship random effect provided
in the package.
}
\keyword{datasets}
